rename pod
update pod on sdk and sample app
add more readme
update version
