+++ RoktSampleApp/README.md
@@ -1,0 +1,30 @@
+# React Native 
+Install ReactNative by running `npm install -g expo-cli`. The full description could be find here https://reactnative.dev/docs/environment-setup
+
+## Install the Rokt widget
+go to *../Rokt.Widget* directory and run `npm install`
+if there is any changes, run `npm pack`
+
+## NPM install on Sample app
+go to *RoktSampleApp* directory and run `npm install`
+
+## Start with Haul
+go to *RoktSampleApp* directory and run `npx haul start`
+
+
+# iOS
+
+## Install the pod
+go to *RoktSampleApp/iOS* directory and run `pod install` (if pod install failed, rund `pod update` to get the latest libraries)
+
+## run iOS
+go to *RoktSampleApp* directory and run `npx react-native run-ios` 
+
+# Android
+
+## run Android
+go to *RoktSampleApp* directory and run `npx react-native run-android` 
+
+
+# Copyright
+Copyright 2020 Rokt Pte Ltd Licensed under the Rokt Software Development Kit (SDK) Terms of Use Version 2.0 (the "License"); You may not use this file except in compliance with the License. You may obtain a copy of the License at https://rokt.com/sdk-license-2-0/
